Initial Commit
I'm dumping this off of my laptop in hopes of preserving the code.  There are a lot of other things in here that I used in my environment to test with.  It is unlikely to build in this state, but the primary purpose of this check-in is to preserve the code.  In a subsequent release I'd like to get this streamlined and working perfectly--perhaps even using docker-compose files to really simplify things.